requirements.txt
Flask==2.2.5
Flask-RESTful
Flask-SQLAlchemy
SQLAlchemy-serializer
flask-cors
psycopg2-binary
python-dotenv
PyJWT
requests

setuptools
flasgger

environs==9.5.0
Werkzeug==2.2.2

pytest
coverage
pytest-cov